requirements.txt
aiosqlite==0.19.0
fastapi==0.108.0
Jinja2==3.1.2
pre-commit==3.6.0
pylibsrtp==0.9.1
python-multipart==0.0.6
shapely==2.0.2
uvicorn==0.25.0
vietocr==0.3.12
yolov5==7.0.13
numpy==1.26.3
opencv-python==4.7.0.72
Pillow==9.5.0
PyYAML==6.0.1
requests==2.31.0
scipy==1.11.4
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
qreader==3.12